Project/fithnity/requirements.txt
Django
djangorestframework
djongo
psycopg2-binary
gunicorn
requests
django-countries
django-phonenumber-field
phonenumbers
djoser
PyJWT
django-cors-headers
drf_yasg